Fix sign-compare build error in MENU_AM modulation string bounds check
Cast int32_t value to unsigned int before comparing with ARRAY_SIZE()
to resolve -Werror=sign-compare.

Agent-Logs-Url: https://github.com/jewelux/quansheng-K5-talking-ardf-rx/sessions/3887675a-b310-4837-a741-5f1a46c7db1a

+++ firmware-source/app/menu.c
@@ -434,7 +434,7 @@ static void MENU_PlayValueVoice(const uint8_t menu_id, const int32_t value)
 			break;
 
 		case MENU_AM:
-			if (value >= 0 && value < ARRAY_SIZE(gModulationStr))
+			if (value >= 0 && (unsigned int)value < ARRAY_SIZE(gModulationStr))
 				MENU_PlayMorseString(gModulationStr[value]);
 			break;
